+++ cmd/validate_test.go
@@ -3,9 +3,8 @@ package cmd
 import (
 	"bytes"
 	"errors"
-	"github.com/gatecheckdev/gatecheck/pkg/artifact/grype"
-	"github.com/gatecheckdev/gatecheck/pkg/config"
 	"github.com/gatecheckdev/gatecheck/pkg/exporter/defectDojo"
+	"github.com/gatecheckdev/gatecheck/pkg/gatecheck"
 	"io"
 	"os"
 	"path"
@@ -52,7 +51,7 @@ func TestValidateCmd(t *testing.T) {
 	})
 
 	t.Run("validation", func(t *testing.T) {
-		c := config.NewConfig("Test Project")
+		c := gatecheck.NewConfig("Test Project")
 		tempConfigFilename := path.Join(os.TempDir(), "gatecheck.yaml")
 
 		if err := OpenAndEncode(tempConfigFilename, YAML, c); err != nil {
@@ -69,7 +68,7 @@ func TestValidateCmd(t *testing.T) {
 	t.Run("report-file-access", func(t *testing.T) {
 		command.SetArgs([]string{"validate", "-c", CopyToTemp(t, TestConfigFilename),
 			"-r", CreateMockFile(t, NoPermissions)})
-		
+
 		if err := command.Execute(); errors.Is(err, ErrorFileAccess) != true {
 			t.Fatalf("Expected file access error, got %v", err)
 		}
@@ -78,19 +77,6 @@ func TestValidateCmd(t *testing.T) {
 }
 
 // Mock Functions
-
-func ConfigFile(t *testing.T) string {
-	c := config.NewConfig("test project")
-	c.Grype = *grype.NewConfig(10)
-	c.Grype.Low = 100
-	c.Grype.Negligible = -1
-	c.Grype.Unknown = -1
-	fPath := path.Join(t.TempDir(), "gatecheck-custom.yaml")
-	if err := OpenAndEncode(fPath, YAML, c); err != nil {
-		t.Fatal(err)
-	}
-	return fPath
-}
 
 func CopyToTemp(t *testing.T, src string) string {
 
@@ -109,43 +95,4 @@ func CopyToTemp(t *testing.T, src string) string {
 		t.Fatal(err)
 	}
 	return fPath
-}
-
-func GrypeTestCopy(t *testing.T, grypeFile io.ReadCloser) string {
-	// Create a temp copy of the grype report
-
-	tempGrypeFilename := path.Join(t.TempDir(), "grype-report.json")
-	tempGrypeFile, _ := os.Create(tempGrypeFilename)
-
-	if _, err := io.Copy(tempGrypeFile, grypeFile); err != nil {
-		t.Fatal(err)
-	}
-	if err := tempGrypeFile.Close(); err != nil {
-		t.Fatal(err)
-	}
-	if err := grypeFile.Close(); err != nil {
-		t.Fatal(err)
-	}
-
-	return tempGrypeFilename
-
-}
-
-func ReportTestCopy(t *testing.T, reportFile io.ReadCloser) string {
-	// Create a temp copy of the grype report
-
-	tempReportFilename := path.Join(t.TempDir(), "gatecheck-report.json")
-	tempReportFile, _ := os.Create(tempReportFilename)
-
-	if _, err := io.Copy(tempReportFile, reportFile); err != nil {
-		t.Fatal(err)
-	}
-	if err := tempReportFile.Close(); err != nil {
-		t.Fatal(err)
-	}
-	if err := reportFile.Close(); err != nil {
-		t.Fatal(err)
-	}
-
-	return tempReportFilename
 }