+++ pkg/gatecheck/validator_test.go
@@ -1,30 +1,27 @@
-package validator_test
+package gatecheck
 
 import (
-	"github.com/gatecheckdev/gatecheck/pkg/config"
-	"github.com/gatecheckdev/gatecheck/pkg/report"
-	"github.com/gatecheckdev/gatecheck/pkg/validator"
 	"testing"
 )
 
 func TestStdValidator_Validate(t *testing.T) {
-	r := report.NewReport("Test Report")
+	r := NewReport("Test Report")
 	r.Artifacts.Grype.Critical.Found = 20
 	r.Artifacts.Grype.High.Found = 22
 	r.Artifacts.Grype.Medium.Found = 113
 
 	t.Run("All Vulnerabilities allowed", func(t *testing.T) {
 		// All vulnerabilities should be allowed
-		c := config.NewConfig("Test Project")
+		c := NewConfig("Test Project")
 
 		r = r.WithConfig(c)
-		if err := validator.NewStdValidator(*r).Validate(); err != nil {
+		if err := NewStdValidator(*r).Validate(); err != nil {
 			t.Fatal(err)
 		}
 	})
 
 	t.Run("Some Allowed", func(t *testing.T) {
-		c := config.NewConfig("Test Project")
+		c := NewConfig("Test Project")
 		c.Grype.Critical = 0
 		c.Grype.High = 1
 		c.Grype.Medium = 2
@@ -33,7 +30,7 @@ func TestStdValidator_Validate(t *testing.T) {
 		c.Grype.Negligible = 0
 		r = r.WithConfig(c)
 
-		err := validator.NewStdValidator(*r).Validate()
+		err := NewStdValidator(*r).Validate()
 		t.Log(err)
 
 		if err == nil {
@@ -42,7 +39,7 @@ func TestStdValidator_Validate(t *testing.T) {
 	})
 
 	t.Run("No vulnerabilities allowed", func(t *testing.T) {
-		c := config.NewConfig("Test Project")
+		c := NewConfig("Test Project")
 		c.Grype.Critical = 0
 		c.Grype.High = 0
 		c.Grype.Medium = 0
@@ -51,7 +48,7 @@ func TestStdValidator_Validate(t *testing.T) {
 		c.Grype.Negligible = 0
 		r = r.WithConfig(c)
 
-		err := validator.NewStdValidator(*r).Validate()
+		err := NewStdValidator(*r).Validate()
 		t.Log(err)
 
 		if err == nil {